+++ mc-java-rejection/src/main/kotlin/io/spine/tools/mc/java/rejection/NoArgMethod.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023, TeamDev. All rights reserved.
+ * Copyright 2022, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,12 +23,26 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-
-package io.spine.tools.mc.java.rejection.v2
+package io.spine.tools.mc.java.rejection
 
 import com.google.errorprone.annotations.Immutable
+import io.spine.tools.java.code.Method
 
+/**
+ * A reference to a method with no arguments.
+ */
 @Immutable
-public data class CodegenContext(
-    public val typeSystem: TypeSystem
-)
+internal class NoArgMethod(methodName: String) : Method(methodName) {
+
+    fun name(): String {
+        return value()
+    }
+
+    fun signature(): String {
+        return value() + "()"
+    }
+
+    companion object {
+        private const val serialVersionUID = 0L
+    }
+}
