+++ mc-java-rejection/src/main/kotlin/io/spine/tools/mc/java/rejection/RejectionRenderer.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.mc.java.rejection.v2
+package io.spine.tools.mc.java.rejection
 
 import com.google.protobuf.BoolValue
 import com.google.protobuf.StringValue
