+++ mc-java-rejection/src/test/kotlin/io/spine/tools/mc/java/rejection/RejectionCodegenSpec.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022, TeamDev. All rights reserved.
+ * Copyright 2023, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,24 +23,21 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package io.spine.tools.mc.java.rejection.gradle
+
+package io.spine.tools.mc.java.rejection
 
 import io.kotest.matchers.shouldBe
 import io.spine.testing.TempDir
 import io.spine.tools.code.SourceSetName
-import io.spine.tools.code.SourceSetName.Companion.main
-import io.spine.tools.code.SourceSetName.Companion.test
-import io.spine.tools.fs.DirectoryName.generated
-import io.spine.tools.fs.DirectoryName.java
+import io.spine.tools.fs.DirectoryName
 import io.spine.tools.gradle.task.JavaTaskName
 import io.spine.tools.gradle.testing.GradleProject
-import io.spine.tools.gradle.testing.GradleProject.Companion.setupAt
 import io.spine.tools.resolve
 import java.io.File
 import java.nio.file.Files
 import java.nio.file.Path
 import kotlin.io.path.isDirectory
-import org.junit.jupiter.api.Assertions.assertTrue
+import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Nested
@@ -57,7 +54,7 @@ internal class RejectionCodegenSpec {
         @JvmStatic
         fun generateRejections() {
             val projectDir = TempDir.forClass(RejectionCodegenSpec::class.java)
-            val project: GradleProject = setupAt(projectDir)
+            val project: GradleProject = GradleProject.setupAt(projectDir)
                 .fromResources("rejection-codegen-test")
                 .copyBuildSrc()
                  /* Uncomment the following line to be able to debug the build.
@@ -75,14 +72,14 @@ internal class RejectionCodegenSpec {
     }
 
     private fun generatedRoot(sourceSetName: SourceSetName): Path =
-        moduleDir.toPath().resolve(generated).resolve(sourceSetName.value)
+        moduleDir.toPath().resolve(DirectoryName.generated).resolve(sourceSetName.value)
 
-    private fun targetMainDir(): Path = generatedRoot(main)
+    private fun targetMainDir(): Path = generatedRoot(SourceSetName.main)
 
-    private fun targetTestDir(): Path = generatedRoot(test)
+    private fun targetTestDir(): Path = generatedRoot(SourceSetName.test)
 
     private fun assertExists(path: Path) =
-        assertTrue(Files.exists(path)) { "The path `$path` is expected to exist." }
+        Assertions.assertTrue(Files.exists(path)) { "The path `$path` is expected to exist." }
 
     private fun assertJavaFileExists(packageDir: Path, typeName: String) {
         val file = packageDir.resolve("$typeName.java")
@@ -95,7 +92,7 @@ internal class RejectionCodegenSpec {
 
         @Test
         fun `'main' source set`() {
-            val mainSpine = targetMainDir().resolve(java)
+            val mainSpine = targetMainDir().resolve(DirectoryName.java)
             assertExists(mainSpine)
             mainSpine.isDirectory() shouldBe true
             mainSpine.containsJavaFiles() shouldBe true
@@ -103,7 +100,7 @@ internal class RejectionCodegenSpec {
 
         @Test
         fun `'test' source set`() {
-            val testSpine = targetTestDir().resolve(java)
+            val testSpine = targetTestDir().resolve(DirectoryName.java)
             assertExists(testSpine)
             testSpine.isDirectory() shouldBe true
             testSpine.containsJavaFiles() shouldBe true
@@ -111,7 +108,9 @@ internal class RejectionCodegenSpec {
 
         @Test
         fun `'testFixtures' source set`() {
-            val testFixturesSpine = generatedRoot(SourceSetName("testFixtures")).resolve(java)
+            val testFixturesSpine = generatedRoot(SourceSetName("testFixtures")).resolve(
+                DirectoryName.java
+            )
             assertExists(testFixturesSpine)
             testFixturesSpine.isDirectory() shouldBe true
             testFixturesSpine.containsJavaFiles() shouldBe true
@@ -125,7 +124,7 @@ internal class RejectionCodegenSpec {
         @Test
         fun `for 'main' source set`() {
             // As defined in `resources/.../main_rejections.proto`.
-            val packageDir = targetMainDir().resolve(java).resolve("io/spine/sample/rejections")
+            val packageDir = targetMainDir().resolve(DirectoryName.java).resolve("io/spine/sample/rejections")
             assertExists(packageDir)
 
             // As defined in `resources/.../main_rejections.proto`.
@@ -140,7 +139,7 @@ internal class RejectionCodegenSpec {
         @Test
         fun `for 'test' source set`() {
             // As defined in `resources/.../test_rejections.proto`.
-            val packageDir = targetTestDir().resolve(java).resolve("io/spine/sample/rejections")
+            val packageDir = targetTestDir().resolve(DirectoryName.java).resolve("io/spine/sample/rejections")
             assertExists(packageDir)
 
             // As defined in `resources/.../test_rejections.proto`.
