+++ mc-java-rejection/src/test/kotlin/io/spine/tools/mc/java/rejection/RejectionJavadocSpec.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022, TeamDev. All rights reserved.
+ * Copyright 2023, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,21 +23,15 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package io.spine.tools.mc.java.rejection.gradle
 
-import com.google.common.truth.Truth.assertThat
+package io.spine.tools.mc.java.rejection
+
+import com.google.common.truth.Truth
 import io.spine.code.java.SimpleClassName
 import io.spine.testing.TempDir
 import io.spine.tools.gradle.testing.GradleProject
-import io.spine.tools.gradle.testing.GradleProject.Companion.setupAt
 import io.spine.tools.java.code.BuilderSpec
-import io.spine.tools.mc.java.gradle.McJavaTaskName.Companion.launchProtoData
-import io.spine.tools.mc.java.rejection.gradle.TestEnv.expectedBuilderClassComment
-import io.spine.tools.mc.java.rejection.gradle.TestEnv.expectedClassComment
-import io.spine.tools.mc.java.rejection.gradle.TestEnv.expectedFirstFieldComment
-import io.spine.tools.mc.java.rejection.gradle.TestEnv.expectedSecondFieldComment
-import io.spine.tools.mc.java.rejection.gradle.TestEnv.rejectionFileContent
-import io.spine.tools.mc.java.rejection.gradle.TestEnv.rejectionsJavadocThrowableSource
+import io.spine.tools.mc.java.gradle.McJavaTaskName
 import org.jboss.forge.roaster.Roaster
 import org.jboss.forge.roaster.model.source.JavaClassSource
 import org.jboss.forge.roaster.model.source.JavaDocCapableSource
@@ -57,12 +51,12 @@ internal class RejectionJavadocSpec {
         @JvmStatic
         fun generateSources() {
             val projectDir = TempDir.forClass(RejectionJavadocSpec::class.java)
-            val project: GradleProject = setupAt(projectDir)
+            val project: GradleProject = GradleProject.setupAt(projectDir)
                 .copyBuildSrc()
                 .fromResources("rejection-javadoc-test") // Provides `build.gradle.kts`
                 .addFile("src/main/proto/javadoc_rejections.proto", rejectionFileContent())
                 .create()
-            project.executeTask(launchProtoData)
+            project.executeTask(McJavaTaskName.launchProtoData)
             val generatedFile = rejectionsJavadocThrowableSource(projectDir.toPath()).toFile()
             generatedSource = Roaster.parse(
                 JavaClassSource::class.java, generatedFile
@@ -109,7 +103,7 @@ internal class RejectionJavadocSpec {
 
     private fun assertDoc(expectedText: String, source: JavaDocCapableSource<*>) {
         val javadoc = source.javaDoc
-        assertThat(javadoc.fullText)
+        Truth.assertThat(javadoc.fullText)
             .isEqualTo(expectedText)
     }
